requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.16
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
